Split shell file validation tests
